requirements.txt
Django>=5.0,<6.0
python-decouple>=3.8
djangorestframework>=3.14
django-cors-headers>=4.0
